different user redirect home

+++ app/Http/Controllers/TasksController.php
@@ -78,8 +78,15 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task=Task::findOrFail($id);
-        return view('tasks.show',['task'=>$task,]);
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+            $task=Task::findOrFail($id);
+            return view('tasks.show',['task'=>$task,]);
+        }else{
+            return redirect('/');            
+        }
+            return redirect('/');
     }
 
     /**
@@ -90,8 +97,15 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task=Task::findOrFail($id);
-        return view('tasks.edit',['task'=>$task,]);
+        $task = \App\Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+            $task=Task::findOrFail($id);
+            return view('tasks.edit',['task'=>$task,]);
+        }else{
+            return redirect('/');
+        }
+        return redirect('/');
+
     }
 
     /**
@@ -107,11 +121,14 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             'content' => 'required|max:255',
         ]);
-        
-        $task=Task::findOrFail($id);
-        $task->status = $request->status;
-        $task->content=$request->content;
-        $task->save();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+            $task=Task::findOrFail($id);
+            $task->status = $request->status;
+            $task->content=$request->content;
+            $task->save();
+        }        
         return redirect('/');
     }
 
@@ -123,8 +140,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task=Task::findOrFail($id);
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 }
